requirements.txt
asyncio==3.4.3
loguru==0.7.2
aiofiles==23.2.1
aiohttp==3.8.6
pyyaml==6.0
requests==2.31.0
uvicorn==0.23.2
nats-py==2.4.0
aioprometheus[aiohttp,binary]==23.3.0
